emitter.rs back under the 800-line budget: fn-ref and map-literal lowering split to emitter_values.rs
